手作業で自動生成とか組んだ
