+++ biome.json
@@ -7,7 +7,12 @@
   },
   "files": {
     "ignoreUnknown": false,
-    "ignore": ["styled-system", "src-tauri", "src/api/generated"]
+    "ignore": [
+      "styled-system",
+      "src-tauri",
+      "src/api/generated",
+      "src/routeTree.gen.ts"
+    ]
   },
   "formatter": {
     "enabled": true,
