+++ docs/PNR.md
@@ -29,6 +29,3 @@ The shape is intentionally simplified to keep the demo small. A real implementat
 - **PNRs list** – shows stored PNRs and links to create or search records.
 - **New PNR** – form to enter record locator, passengers and flight segments.
 - **Search PNR** – filter stored PNRs by record locator.
-
-The pages live under `src/pages/pnrs` and are linked to routes generated with `createFileRoute` under `src/routes`.
-
