+++ package.json
@@ -18,7 +18,8 @@
     "gen:api": "npx openapi-typescript http://aeromate-backend.fuminori.workers.dev/doc -o ./src/api/generated/types.ts"
   },
   "dependencies": {
-    "@tanstack/react-router": "^1.120.15",
+    "@tanstack/react-router": "^1.120.16",
+    "@tanstack/router-plugin": "^1.120.16",
     "@tauri-apps/api": "^2",
     "@tauri-apps/plugin-opener": "^2",
     "@types/js-cookie": "^3.0.6",
@@ -26,7 +27,8 @@
     "react": "^18.3.1",
     "react-aria-components": "^1.9.0",
     "react-dom": "^18.3.1",
-    "sass": "^1.89.1"
+    "sass": "^1.89.1",
+    "vite": "^6.3.5"
   },
   "devDependencies": {
     "@biomejs/biome": "^1.9.4",
@@ -43,8 +45,7 @@
     "openapi-typescript": "^7.8.0",
     "pdf417-generator": "^1.0.5",
     "react-hook-form": "^7.56.4",
-    "typescript": "~5.6.2",
-    "vite": "^6.0.3"
+    "typescript": "~5.6.2"
   },
   "volta": {
     "node": "22.16.0",
