--- src/pages/home/index.tsx
@@ -1,20 +1,0 @@
-import { css } from "../../../styled-system/css";
-import { MainLayout } from "../../layouts/MainLayout";
-
-function App() {
-  return (
-    <MainLayout>
-      <h1
-        className={css({
-          fontSize: "1.7rem",
-          fontWeight: "700",
-          color: "#333",
-        })}
-      >
-        Welcome
-      </h1>
-    </MainLayout>
-  );
-}
-
-export default App;
