+++ tsconfig.json
@@ -21,5 +21,6 @@
     "noFallthroughCasesInSwitch": true
   },
   "include": ["src"],
+  "exclude": ["src/routeTree.gen.ts"],
   "references": [{ "path": "./tsconfig.node.json" }]
 }
